# [model] new API for efficient cache control