--- a/wenet/transformer/convolution.py
+++ b/wenet/transformer/convolution.py
@@ -5,7 +5,7 @@
 # Author: di.wu@mobvoi.com (DI WU)
 """ConvolutionModule definition."""
 
-from typing import Optional, Tuple
+from typing import Tuple
 
 import torch
 from torch import nn
@@ -81,15 +81,17 @@ class ConvolutionModule(nn.Module):
     def forward(
         self,
         x: torch.Tensor,
-        mask_pad: Optional[torch.Tensor] = None,
-        cache: Optional[torch.Tensor] = None,
+        mask_pad: torch.Tensor = torch.ones((0, 0, 0), dtype=torch.bool),
+        cache: torch.Tensor = torch.zeros((0, 0, 0)),
     ) -> Tuple[torch.Tensor, torch.Tensor]:
         """Compute convolution module.
         Args:
             x (torch.Tensor): Input tensor (#batch, time, channels).
-            mask_pad (torch.Tensor): used for batch padding (#batch, 1, time)
+            mask_pad (torch.Tensor): used for batch padding (#batch, 1, time),
+                (0, 0, 0) means fake mask.
             cache (torch.Tensor): left context cache, it is only
-                used in causal convolution
+                used in causal convolution (#batch, channels, cache_t),
+                (0, 0, 0) meas fake cache.
         Returns:
             torch.Tensor: Output tensor (#batch, time, channels).
         """
@@ -97,15 +99,15 @@ class ConvolutionModule(nn.Module):
         x = x.transpose(1, 2)  # (#batch, channels, time)
 
         # mask batch padding
-        if mask_pad is not None:
+        if mask_pad.size(2) > 0:  # time > 0
             x.masked_fill_(~mask_pad, 0.0)
 
         if self.lorder > 0:
-            if cache is None:
+            if cache.size(2) == 0:  # cache_t == 0
                 x = nn.functional.pad(x, (self.lorder, 0), 'constant', 0.0)
             else:
-                assert cache.size(0) == x.size(0)
-                assert cache.size(1) == x.size(1)
+                assert cache.size(0) == x.size(0)  # equal batch
+                assert cache.size(1) == x.size(1)  # equal channel
                 x = torch.cat((cache, x), dim=2)
             assert (x.size(2) > self.lorder)
             new_cache = x[:, :, -self.lorder:]
@@ -113,7 +115,7 @@ class ConvolutionModule(nn.Module):
             # It's better we just return None if no cache is requried,
             # However, for JIT export, here we just fake one tensor instead of
             # None.
-            new_cache = torch.tensor([0.0], dtype=x.dtype, device=x.device)
+            new_cache = torch.zeros((0, 0, 0), dtype=x.dtype, device=x.device)
 
         # GLU mechanism
         x = self.pointwise_conv1(x)  # (batch, 2*channel, dim)
@@ -128,7 +130,7 @@ class ConvolutionModule(nn.Module):
             x = x.transpose(1, 2)
         x = self.pointwise_conv2(x)
         # mask batch padding
-        if mask_pad is not None:
+        if mask_pad.size(2) > 0:  # time > 0
             x.masked_fill_(~mask_pad, 0.0)
 
         return x.transpose(1, 2), new_cache