# Benchmark(cpp)
Model: Aishell-1 u2++ Conformer (with **dynamic quantization**)
Config: num_bins 80, ctc_weight 0.3, rescoring_weight 1.0, reverse_weight 0.5
Test-set: Aishell-1 test-set, 7176 sentences
| API   | chunk_size | num_left_chunks | RTF(OMP_NUM_THREADS=1) | WER(int8) | WER(float32) |
|-|-|-|-|-|-|
| old   | -1 | -1  |  0.05871  |  4.89 %  |  4.68 %  |
|         | 16 | -1  | 0.1057 |  5.27 %    | 5.07 % |
|         | 16 | 4   | 0.08996  |  5.73 % | 5.60 %  |
| new | -1  | -1  |  0.05859 | 4.89 %  | 4.68 % |
|         | 16 | -1  | 0.07453 | 5.19 %  | 5.07 % |
|         | 16 | 4   | 0.07064  | 5.68 %  | 5.60 % |

# Benchmark(python)
Model: Aishell-1 u2++ Conformer
Config: num_bins 80, ctc_weight 0.3, rescoring_weight 1.0, reverse_weight 0.5
Test-set: Aishell-1 test-set, 7176 sentences
| API   | chunk_size | num_left_chunks | WER(float32) |
|-|-|-|-|
| old   | -1 | -1  |  4.68 %  |
|         | 16 | -1  | 5.07 % |
|         | 16 | 4   | 5.60 %  |
| new | -1  | -1  | 4.68 % |
|         | 16 | -1  | 5.07 % |
|         | 16 | 4   | 5.60 % |

# TODO
- [x] cpp benchmark tests using libtorch and `forward_chunk`
- [x] python benchmark tests using PyTorch and `forward_chunk_by_chunk`
- [x] check training behaviour, should be equal to old API.

# Q&A
1. Why new API is faster than old API ?
 
![cache-control](https://user-images.githubusercontent.com/13466943/161366079-c4ab1f59-f35b-4ead-8048-3014de0fe806.png)

2. Why new API is more accurate than old API for **dynamically quantized model**?

In dynamic quantization, the scale factor for activations is determined dynamically based on the data range observed at runtime, this means that different input results in different scale factor. In old API, the input of `ffn_macaron` contains 5 chunks (4 cache + 1 data), The output of `ffn_macaron` corresponding to the cache chunk might be changed due to the recalculation of scale point which is observed on both cache and data. However, in new API, cache will never be changed.