@Slyne 2022-04-02
Thanks! The pic is quite clear. Haven't tested it yet. Will come back soon.