@robin1001 2022-04-02
@Slyne , please refactor your gpu onnx according to the new interface. And I think it's better to rename export_onnx.py to export_onnx_gpu.py since we will add cpu onnx support in a few weeks.